Image alignment
I  took the section tag off the images and aligned it manually .  There is also so imformation missing from this github contenct versus what's  on the actually website.  I added the Welcome to our site text and aligned manually as well.   I'm adding Css info to the Welcome text on my  stylesheet page.   I reccommended the pic on this page be sized on your end to a smaller size  and please done center by coding. I would put the measurements in here but  wasn't sure how to view in real time.  I'll leave to you as well.   Manually center it by viewing the page and determine whats a good spot to put it.