Harrison's files
I've uploaded two files. They are design in nature. I suggest to at least try the coding implemted in on the site and see if you like it.   The design is simple and not busy.  A simple background to take away all of the white.  And a few flares of blue to add a simple design throughout continuity of site.